+++ app/src/main/java/com/example/slacks_lottoevent/EventsHomeActivity.java
@@ -1,15 +1,26 @@
 package com.example.slacks_lottoevent;
 
+import android.app.Activity;
 import android.app.AlertDialog;
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
+import android.os.Build;
 import android.os.Bundle;
+import android.provider.Settings;
+import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+import androidx.core.app.NotificationCompat;
+import androidx.core.app.NotificationManagerCompat;
 import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
 import androidx.navigation.ui.AppBarConfiguration;
@@ -17,6 +28,8 @@ import androidx.navigation.ui.AppBarConfiguration;
 import com.example.slacks_lottoevent.databinding.ActivityEventsHomeBinding;
 import com.google.android.material.appbar.MaterialToolbar;
 import com.google.android.material.tabs.TabLayout;
+import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.FirebaseFirestore;
 
 
 /**
@@ -27,6 +40,11 @@ public class EventsHomeActivity extends AppCompatActivity {
     private ActivityEventsHomeBinding binding;
     private AppBarConfiguration appBarConfiguration;
     private MaterialToolbar toolbar;
+
+    private NotificationHelper notificationHelper;
+
+    private static final int NOTIFICATION_PERMISSION_REQUEST_CODE = 1;
+
 
     /**
      * onCreate method for the EventsHomeActivity.
@@ -48,6 +66,12 @@ public class EventsHomeActivity extends AppCompatActivity {
 
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment_content_events_home); // Get the navigation controller
         appBarConfiguration = new AppBarConfiguration.Builder(navController.getGraph()).build(); // Build the app bar configuration
+
+
+        createNotificationChannel();
+        notificationHelper = new NotificationHelper(this);
+        checkAndRequestNotificationPermission();
+
 
         eventsTabs.addOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
             @Override
@@ -166,4 +190,81 @@ public class EventsHomeActivity extends AppCompatActivity {
         }
         return super.onOptionsItemSelected(item);
     }
+
+    private void createNotificationChannel() {
+        // Create the NotificationChannel, but only on API 26+ because
+        // the NotificationChannel class is not in the Support Library.
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            // Use the appropriate context method
+            String channelId = getString(R.string.channel_id); // ensure channel ID is properly defined in strings.xml
+            CharSequence name = getString(R.string.channel_name); // name of the channel
+            String description = getString(R.string.channel_description); // description of the channel
+            int importance = NotificationManager.IMPORTANCE_DEFAULT; // adjust as necessary
+
+            // Create the NotificationChannel
+            NotificationChannel channel = new NotificationChannel(channelId, name, importance);
+            channel.setDescription(description);
+
+            // Register the channel with the system
+            NotificationManager notificationManager = getSystemService(NotificationManager.class);
+            if (notificationManager != null) {
+                notificationManager.createNotificationChannel(channel);
+            }
+        }
+    }
+
+    private void grabbingNotifications(String deviceId){
+        FirebaseFirestore db = FirebaseFirestore.getInstance();
+        db.collection("notifications")
+                .whereEqualTo("userId", deviceId)  // Match documents where userId is "user1"
+                .get()
+                .addOnSuccessListener(queryDocumentSnapshots -> {
+                    for (DocumentSnapshot document : queryDocumentSnapshots) {
+                        // doc3 will be included in the results, along with other matching documents
+                        String title = document.getString("title");
+                        String messageContent = document.getString("message");
+                        notificationHelper.sendNotifications(deviceId, title, messageContent);
+
+                    }
+                })
+                .addOnFailureListener(e -> {
+                    Log.e("Firestore", "Error fetching notifications", e);
+                });
+    }
+
+    private void checkAndRequestNotificationPermission() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            if (ActivityCompat.checkSelfPermission(this, android.Manifest.permission.POST_NOTIFICATIONS) != PackageManager.PERMISSION_GRANTED) {
+                ActivityCompat.requestPermissions(this,
+                        new String[]{android.Manifest.permission.POST_NOTIFICATIONS},
+                        NOTIFICATION_PERMISSION_REQUEST_CODE);
+            } else {
+                startFetchingNotifications();
+            }
+        } else {
+            startFetchingNotifications();
+        }
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        if (requestCode == NOTIFICATION_PERMISSION_REQUEST_CODE) {
+            if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                startFetchingNotifications();
+            } else {
+                Log.d("EventsHomeActivity", "Notification permission denied.");
+            }
+        }
+    }
+
+    private void startFetchingNotifications() {
+        String deviceId = Settings.Secure.getString(getContentResolver(), Settings.Secure.ANDROID_ID);
+        grabbingNotifications(deviceId);
+
+        Notifications notification = new Notifications();
+        notification.removeNotifications(deviceId);
+    }
+
+
 }
